+++ CHANGELOG.md
@@ -474,6 +474,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#1135 DateTimes now properly get time zones on derserialization.
 
+* \#1031 Mongoid now serializes values in Array fields to their proper
+  Mongo-friendly values when possible.
+
 * \#685 Attempting to use versioning with embedded documents will now
   raise a proper error alerting the developer this is not allowed.